backend/requirements.txt
babel>=2.14.0,<3.0
Django>=4.2,<5.0
boto3>=1.34,<2
celery>=5.2.6,<6.0
django-ckeditor-5>=0.2,<1.0
# django_ckeditor>=6.2,<7.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
django-oauth-toolkit>=2.3,<3.0
django-storages>=1.14.2,<2.0
djangorestframework>=3.14,<4.0
djangorestframework-gis>=0.1,<1.0
django-daterangefilter>=1.0,<2.0
django-material-admin>=1.7,<2.0
django-pandas>=0.6.6,<1.0
django-redis>=5.2,<6.0
django-admin-rangefilter>=0.12,<1.0
mysqlclient>=1.3.14,<2.0
fpdf2>=2.6.0,<3.0
gunicorn>=20.1,<21.0
matplotlib>=3.8,<4.0
pandas>=2.2,<3.0
psycopg2>=2.9.9,<3
redis>=3.2,<4.0
requests>=2.31,<3
sqlalchemy>=1.3,<2.0
xlrd>=2.0,<3.0
xlsxwriter>=3.2,<4.0